refactor(tests): consolidate image-related fixtures into conftest.py for better test organization and and re-use
